+++ SpellActivationOverlay.lua
@@ -11,6 +11,7 @@ local longSide = 256 * sizeScale;
 local shortSide = 128 * sizeScale;
 local combatOverlayFactor = 2;
 local useTimer = true;
+local useSound = false;
 
 function SpellActivationOverlay_OnLoad(self)
 	SAO.Frame = self;
@@ -29,6 +30,9 @@ function SpellActivationOverlay_OnLoad(self)
 
 	self.useTimer = true;
 	SpellActivationOverlay_OnChangeTimerVisibility(self);
+
+	self.useSound = false;
+	SpellActivationOverlay_OnChangeSoundToggle(self);
 
 	local className, classFile, classId = UnitClass("player");
 	local class = SAO.Class[classFile];
@@ -90,11 +94,45 @@ end
 function SpellActivationOverlay_OnChangeTimerVisibility(self)
 	SAO:Trace(Module, "SpellActivationOverlay_OnChangeTimerVisibility");
 
+	if useTimer == self.useTimer then
+		return;
+	end
+
 	useTimer = self.useTimer;
+
 	for _, overlayList in pairs(self.overlaysInUse) do
 		for i=1, #overlayList do
 			local overlay = overlayList[i];
 			overlay.mask:SetShown(useTimer);
+		end
+	end
+end
+
+function SpellActivationOverlay_OnChangeSoundToggle(self)
+	SAO:Trace(Module, "SpellActivationOverlay_OnChangeSoundToggle");
+
+	if useSound == self.useSound then
+		return;
+	end
+
+	useSound = self.useSound;
+
+	if useSound then
+		for _, overlayList in pairs(self.overlaysInUse) do
+			if #overlayList > 0 then
+				-- Play generic sound if at least one effect is displayed
+				-- No need to spam players with several effects, because currently there is only one type of sound effect
+				overlayList[1].soundHandle = SAO:PlaySpellAlertSound();
+				break;
+			end
+		end
+	else
+		for _, overlayList in pairs(self.overlaysInUse) do
+			for i=1, #overlayList do
+				local overlay = overlayList[i];
+				SAO:StopSpellAlertSound(overlay.soundHandle);
+				overlay.soundHandle = nil;
+			end
 		end
 	end
 end
@@ -274,7 +312,9 @@ function SpellActivationOverlay_ShowOverlay(self, spellID, texturePath, position
 	overlay.texture:SetVertexColor(r / 255, g / 255, b / 255);
 	
 	overlay.animOut:Stop();	--In case we're in the process of animating this out.
-	PlaySound(SOUNDKIT.UI_POWER_AURA_GENERIC);
+	if useSound then
+		overlay.soundHandle = SAO:PlaySpellAlertSound();
+	end
 	overlay:Show();
 	if ( forcePulsePlay ) then
 		overlay.pulse:Play();
@@ -446,6 +486,10 @@ function SpellActivationOverlayTexture_TerminateOverlay(overlay)
 	overlay.mask.timeoutY:Stop();
 	overlay.combat.animIn:Stop();
 	overlay.combat.animOut:Stop();
+
+	-- Stop playing sound
+	SAO:StopSpellAlertSound(overlay.soundHandle, 1000);
+	overlay.soundHandle = nil;
 
 	-- Hide the overlay and make it available again in the pool for future use
 	overlay.mask:SetScale(1); -- Reset scale, in case a previous animation shrank it to 0.01